+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - MSFT_xNetAdapterAdvancedProperty
  - Created new resource configuring AdvancedProperties for NetAdapter
+- MSFT_xIPAddressOption:
+  - Added a new resource to set the SkipAsSource option for an IP address.
+- MSFT_xWeakHostSend:
+  - Created the new Weak Host Send resource.
+- MSFT_xWeakHostReceive:
+  - Created the new Weak Host Receive resource.
 
 ## 5.3.0.0